Re-org function hasAccessToResource
